Raisa

+++ src/main/java/sri/sysint/sri_starter_back/controller/ProductController.java
@@ -40,10 +40,16 @@ import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 
 import sri.sysint.sri_starter_back.exception.ResourceNotFoundException;
+import sri.sysint.sri_starter_back.model.Pattern;
 import sri.sysint.sri_starter_back.model.Plant;
 import sri.sysint.sri_starter_back.model.Product;
+import sri.sysint.sri_starter_back.model.ProductType;
 import sri.sysint.sri_starter_back.model.Response;
+import sri.sysint.sri_starter_back.model.Size;
 import sri.sysint.sri_starter_back.model.TassSize;
+import sri.sysint.sri_starter_back.repository.PatternRepo;
+import sri.sysint.sri_starter_back.repository.ProductTypeRepo;
+import sri.sysint.sri_starter_back.repository.SizeRepo;
 import sri.sysint.sri_starter_back.service.ProductServiceImpl;
 
 @CrossOrigin(maxAge = 3600)
@@ -54,6 +60,14 @@ public class ProductController {
 	@Autowired
 	private ProductServiceImpl productServiceImpl;
 	
+	@Autowired
+    private ProductTypeRepo productTypeRepo;
+	
+	@Autowired
+    private SizeRepo sizeRepo;
+	
+	@Autowired
+    private PatternRepo patternRepo;
 	@PersistenceContext	
 	private EntityManager em;
 	
@@ -283,33 +297,33 @@ public class ProductController {
 	
 	@PostMapping("/saveProductExcel")
 	public Response saveProductExcelFile(@RequestParam("file") MultipartFile file, final HttpServletRequest req) throws ResourceNotFoundException {
-//	    String header = req.getHeader("Authorization");
-//
-//	    if (header == null || !header.startsWith("Bearer ")) {
-//	        throw new ResourceNotFoundException("JWT token not found or maybe not valid");
-//	    }
-//
-//	    String token = header.replace("Bearer ", "");
-//
-//	    try {
-//	        String user = JWT.require(Algorithm.HMAC512(SECRET.getBytes()))
-//	            .build()
-//	            .verify(token)
-//	            .getSubject();
-//
-//	        if (user != null) {
+	    String header = req.getHeader("Authorization");
+
+	    if (header == null || !header.startsWith("Bearer ")) {
+	        throw new ResourceNotFoundException("JWT token not found or maybe not valid");
+	    }
+
+	    String token = header.replace("Bearer ", "");
+
+	    try {
+	        String user = JWT.require(Algorithm.HMAC512(SECRET.getBytes()))
+	            .build()
+	            .verify(token)
+	            .getSubject();
+
+	        if (user != null) {
 	            if (file.isEmpty()) {
 	                return new Response(new Date(), HttpStatus.BAD_REQUEST.value(), null, "No file uploaded", req.getRequestURI(), null);
 	            }
-	            
+
 	            productServiceImpl.deleteAllProduct();
 
 	            try (InputStream inputStream = file.getInputStream()) {
 	                XSSFWorkbook workbook = new XSSFWorkbook(inputStream);
 	                XSSFSheet sheet = workbook.getSheetAt(0);
-	                
+
 	                List<Product> products = new ArrayList<>();
-	                
+
 	                for (int i = 1; i <= sheet.getLastRowNum(); i++) {
 	                    Row row = sheet.getRow(i);
 	                    if (row == null || isRowEmpty(row)) {
@@ -317,20 +331,41 @@ public class ProductController {
 	                    }
 
 	                    Product product = new Product();
-	                    product.setPART_NUMBER(getBigDecimalFromCell(row.getCell(0))); // Part number
-	                    product.setITEM_CURING(getStringFromCell(row.getCell(1)));
-	                    product.setPATTERN_ID(getBigDecimalFromCell(row.getCell(2)));
-	                    product.setSIZE_ID(getStringFromCell(row.getCell(3)));
-	                    product.setPRODUCT_TYPE_ID(getBigDecimalFromCell(row.getCell(4)));
-	                    product.setDESCRIPTION(getStringFromCell(row.getCell(5)));
-	                    product.setRIM(getBigDecimalFromCell(row.getCell(6)));
-	                    product.setWIB_TUBE(getStringFromCell(row.getCell(7)));
-	                    product.setITEM_ASSY(getStringFromCell(row.getCell(8)));
-	                    product.setITEM_EXT(getStringFromCell(row.getCell(9)));
-	                    product.setEXT_DESCRIPTION(getStringFromCell(row.getCell(10)));
-	                    product.setQTY_PER_RAK(getBigDecimalFromCell(row.getCell(11)));
-	                    product.setUPPER_CONSTANT(getBigDecimalFromCell(row.getCell(12)));
-	                    product.setLOWER_CONSTANT(getBigDecimalFromCell(row.getCell(13)));
+
+	                    product.setPART_NUMBER(getBigDecimalFromCell(row.getCell(1))); 
+	                    product.setITEM_CURING(getStringFromCell(row.getCell(2)));
+
+	                    // PATTERN_NAME -> PATTERN_ID
+	                    String patternName = getStringFromCell(row.getCell(3));
+	                    Optional<Pattern> pattern = patternRepo.findByName(patternName);
+	                    if (pattern.isPresent()) {
+	                        product.setPATTERN_ID(pattern.get().getPATTERN_ID());
+	                    } else {
+	                        throw new ResourceNotFoundException("Pattern not found: " + patternName);
+	                    }
+
+	                    product.setSIZE_ID(getStringFromCell(row.getCell(4)));
+
+
+	                    // CATEGORY -> PRODUCT_TYPE_ID
+	                    String category = getStringFromCell(row.getCell(5));
+	                    Optional<ProductType> productType = productTypeRepo.findByCategory(category);
+	                    if (productType.isPresent()) {
+	                        product.setPRODUCT_TYPE_ID(productType.get().getPRODUCT_TYPE_ID());
+	                    } else {
+	                        throw new ResourceNotFoundException("Product type not found: " + category);
+	                    }
+
+	                    product.setDESCRIPTION(getStringFromCell(row.getCell(6)));
+	                    product.setRIM(getBigDecimalFromCell(row.getCell(7)));
+	                    product.setWIB_TUBE(getStringFromCell(row.getCell(8)));
+	                    product.setITEM_ASSY(getStringFromCell(row.getCell(9)));
+	                    product.setITEM_EXT(getStringFromCell(row.getCell(10)));
+	                    product.setEXT_DESCRIPTION(getStringFromCell(row.getCell(11)));
+	                    product.setQTY_PER_RAK(getBigDecimalFromCell(row.getCell(12)));
+	                    product.setUPPER_CONSTANT(getBigDecimalFromCell(row.getCell(13)));
+	                    product.setLOWER_CONSTANT(getBigDecimalFromCell(row.getCell(14)));
+
 	                    product.setSTATUS(BigDecimal.valueOf(1));
 	                    product.setCREATION_DATE(new Date());
 	                    product.setLAST_UPDATE_DATE(new Date());
@@ -340,20 +375,22 @@ public class ProductController {
 	                }
 
 	                response = new Response(new Date(), HttpStatus.OK.value(), null, "File processed and data saved", req.getRequestURI(), products);
-	                
+
 	            } catch (IOException e) {
 	                response = new Response(new Date(), HttpStatus.INTERNAL_SERVER_ERROR.value(), null, "Error processing file", req.getRequestURI(), null);
 	            }
 
-//	        } else {
-//	            throw new ResourceNotFoundException("User not found");
-//	        }
-//	    } catch (Exception e) {
-//	        throw new ResourceNotFoundException("JWT token is not valid or expired");
-//	    }
+	        } else {
+	            throw new ResourceNotFoundException("User not found");
+	        }
+	    } catch (Exception e) {
+	        throw new ResourceNotFoundException("JWT token is not valid or expired");
+	    }
 
 	    return response;
 	}
+
+
 
 	private boolean isRowEmpty(Row row) {
 	    for (int j = 0; j < row.getLastCellNum(); j++) {
@@ -389,17 +426,7 @@ public class ProductController {
 	    if (cell == null) {
 	        return null;
 	    }
-	    
-	    switch (cell.getCellType()) {
-	        case STRING:
-	            return cell.getStringCellValue();
-	        case NUMERIC:
-	            return String.valueOf(cell.getNumericCellValue());
-	        case BOOLEAN:
-	            return String.valueOf(cell.getBooleanCellValue());
-	        default:
-	            return null; // Handle other cell types as needed
-	    }
+	    return cell.getStringCellValue();
 	}
 	
     @RequestMapping("/exportProductsExcel")
